backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
python-decouple>=3.8
coverage>=7.0,<9.0

